First Commit:
Set up basic logic utilizing a binary search tree and queue to determine possible moves (within boundaries of the board) and print an array of the history of the Knight's moves. Need to resolve bug where the final position is not printing as part of the history and history seems to include positions from other child branches.